requirements.txt
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
Django==3.2
django-admin-honeypot==1.1.0
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
Faker==15.3.1
importlib-metadata==5.0.0
isort==5.10.1
lazy-object-proxy==1.8.0
Markdown==3.4.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pycodestyle==2.9.1
pycrypto==2.6.1
pylint==2.15.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-redis-lock==4.0.0
pytz==2022.6
redis==4.3.4
redis-cache==0.1.5
redis-structures==0.1.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
validate-docbr==1.10.0
vital-tools==0.1.13
wrapt==1.14.1
zipp==3.10.0
